+++ src/controller/CommissionerDiscoveryController.cpp
@@ -122,7 +122,6 @@ void CommissionerDiscoveryController::Ok()
 {
     ChipLogDetail(AppServer, "UX Ok: moving to main thread");
     // need to ensure callback is on main chip thread
-    // chip::DeviceLayer::StackLock lock;
     assertChipStackLockedByCurrentThread();
     DeviceLayer::SystemLayer().StartTimer(System::Clock::Seconds32(0), CallbackOk, this);
 }
